Add a few unit tests and clean up a couple of other things.
- Add a test to verify that `make_move()` and `unmake_move()` work. This
  will probably be useful later.
- Add a test for `decompose_move()`, since `create_move()` has one.
- Move `bitboard_from_pos()` into `bits.rs` and make `rank_to_char()`
  an associated function of `Board`.
- Make `File` and `Rank` a usize instead of u8, as their primary purpose
  is indexing. There's a small chance this comes at a minor performance
  penalty (since they now use all of a register) but that isn't really
  an issue.
- Make `Magic.shift` a u32, since it doesn't change the size of the
  struct and it removes the need to case it as a u32.

+++ src/movegen.rs
@@ -72,12 +72,12 @@ impl Movegen {
             | ray_attacks[Directions::SW][square]
             | ray_attacks[Directions::NW][square];
         rays &=
-            !(Bitboards::FILE_BB[Files::FILE1 as usize]
-              | Bitboards::FILE_BB[Files::FILE8 as usize])
+            !(Bitboards::FILE_BB[Files::FILE1]
+              | Bitboards::FILE_BB[Files::FILE8])
             | Bitboards::FILE_BB[file_of(square) as usize];
         rays &=
-            !(Bitboards::RANK_BB[Ranks::RANK1 as usize]
-              | Bitboards::RANK_BB[Ranks::RANK8 as usize])
+            !(Bitboards::RANK_BB[Ranks::RANK1]
+              | Bitboards::RANK_BB[Ranks::RANK8])
             | Bitboards::RANK_BB[rank_of(square) as usize];
         rays
     }
@@ -94,12 +94,12 @@ impl Movegen {
             | ray_attacks[Directions::S][square]
             | ray_attacks[Directions::W][square];
         rays &=
-            !(Bitboards::FILE_BB[Files::FILE1 as usize]
-              | Bitboards::FILE_BB[Files::FILE8 as usize])
+            !(Bitboards::FILE_BB[Files::FILE1]
+              | Bitboards::FILE_BB[Files::FILE8])
             | Bitboards::FILE_BB[file_of(square) as usize];
         rays &=
-            !(Bitboards::RANK_BB[Ranks::RANK1 as usize]
-              | Bitboards::RANK_BB[Ranks::RANK8 as usize])
+            !(Bitboards::RANK_BB[Ranks::RANK1]
+              | Bitboards::RANK_BB[Ranks::RANK8])
             | Bitboards::RANK_BB[rank_of(square) as usize];
         rays
     }
